Added the option to change input scheme for this processor.
Currently only EDGAR or IITM inputs are enabled. And dummy
emissions need to be created in order to deal with missing
emission species (e.g. IITM has no NH3 emissions, but we
still need an empty NH3_AGR input so that we can create
WRF-Chem emission files for schemes which expect there to be
NH3 emissions).

Also added the wind-blown-dust (WBD) emission category.