Correcting some chemical formulae.
Apparently I don't know the difference between
a 5 and a 6. So the variable names that were C6H12 and C8H14
have been corrected to their proper formulae of C5H12 and
C6H14. There is no change in the output from this, just
an important (but cosmetic) correction of the variable names.