Bug fix - emiss_cat_index@VARIABLE
Corrected the indexing for emissions data used
for direct mapping from emission species (usually CO, not an NMVOC)
in process_nmvoc_emissions_direct. This error was causing
the use of uninitialised data in the "emissions_data" array,
which then led to writing NaN's to the final emission file
(which WRF-Chem can't cope with!).