Setting nc format option to NetCDF4Classic.
Found a problem with the default "Classic" netcdf format,
where creating too many large data arrays at the same time
causes silent failure of the 'filevardef' routine.

The fix is to switch to NetCDF4 or NetCDF4Classic. This option
has been set now at the start of the main script, so that all
files created by the tool will be NETCDF4Classic format.